requirements.txt
click==8.0.3
hydra-core==1.1.1
matplotlib==3.5.1
numpy==1.22.0
omegaconf==2.1.1
pandas==1.3.5
pytest==6.2.5
python-dotenv==0.19.2
scikit_learn==1.0.2
seaborn==0.11.2
setuptools==60.2.0
torch==1.10.1
torchsummary==1.5.1
torchvision==0.11.2
